+++ src/main/java/com/project/zodiac_api/dto/ClientResponseDto.java
@@ -19,33 +19,32 @@ public class ClientResponseDto {
     private LocalTime birthTime;
     private String birthPlace;
 
-     // ── v9：上升 / 天頂四軸資訊（允許 null）────────────────────────────
-    private String ascSign;
-    private Short  ascDegreeNum;
-    private Short  ascMinuteNum;
- 
-    private String mcSign;
-    private Short  mcDegreeNum;
-    private Short  mcMinuteNum;
-    // ─────────────────────────────────────────────────────────────────────
-    
+    // 上升 / 天頂四軸資訊（允許 null）
+    private String  ascSign;
+    private Integer ascDegreeNum;
+    private Integer ascMinuteNum;
+
+    private String  mcSign;
+    private Integer mcDegreeNum;
+    private Integer mcMinuteNum;
+
     private LocalDateTime createdAt;
 
+    /** 單筆查詢用（含 ASC/MC）：GET /api/clients/{id} */
     public static ClientResponseDto from(Client c) {
         ClientResponseDto dto = new ClientResponseDto();
-        dto.id          = c.getId();
-        dto.name        = c.getName();
-        dto.birthDate   = c.getBirthDate();
-        dto.birthTime   = c.getBirthTime();
-        dto.birthPlace  = c.getBirthPlace();
-        // v9：四軸資訊
+        dto.id           = c.getId();
+        dto.name         = c.getName();
+        dto.birthDate    = c.getBirthDate();
+        dto.birthTime    = c.getBirthTime();
+        dto.birthPlace   = c.getBirthPlace();
         dto.ascSign      = c.getAscSign();
         dto.ascDegreeNum = c.getAscDegreeNum();
         dto.ascMinuteNum = c.getAscMinuteNum();
         dto.mcSign       = c.getMcSign();
         dto.mcDegreeNum  = c.getMcDegreeNum();
         dto.mcMinuteNum  = c.getMcMinuteNum();
-        dto.createdAt   = c.getCreatedAt();
+        dto.createdAt    = c.getCreatedAt();
         return dto;
     }
 }
